Big rename, following patterns and conventions from embedded-hal and/or embassy
