+++ ssh-stamp-hal/src/traits/hash.rs
