+++ ssh-stamp-hal/src/traits/network/wifi.rs
