+++ ssh-stamp-hal/src/traits/rng.rs
