+++ ssh-stamp-hal/src/traits/timer.rs
