+++ ssh-stamp-hal/src/traits/uart.rs
